pyproject.toml
[tool.poetry]
name = "eventtom"
version = "0.1.0"
description = ""
authors = ["Max Tepper <FlashwarLP@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg = "^3.2.3"
djangorestframework-simplejwt = "^5.3.1"
pytest = "^8.3.3"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
python-dotenv = "^1.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
